al pedi :poop:

+++ js/subject.js
@@ -1,7 +1,9 @@
+var alertify = require("alertifyjs");
+
 function salir() {
     alertify.error("¡Nos vemos pronto!");
     setTimeout(function () {
-        location.href = "index.html";
+        location.href = "signin.html";
     }, 3000);
 }
 
@@ -23,10 +25,10 @@ var app = new Vue({
             console.log(document.getElementById('text').removeChild(document.getElementById('text').lastChild))
         },
         cupdate: function () {
-            document.getElementById('update').setAttribute("class", "modal")
-
-
-
+            alertify.error("Se canceló la actualización de la información de la materia.");
+            setTimeout(function () {
+                location.href = "subject.html";
+            }, 3000);
         },
         del: function (id) {
             console.log(id)
@@ -81,32 +83,56 @@ var app = new Vue({
                         span2.appendChild(span)
                         span2.setAttribute("class", "button is-danger")
                         span2.addEventListener("click", function () {
-                            p = document.createElement('p')
-                            p.appendChild(document.createTextNode(element['VCH_NOMBRE']))
-                            document.getElementById('delete').setAttribute("class", "is-active modal")
-                            document.getElementById('text').appendChild(p)
-                            document.getElementById('delete-btn').addEventListener("click", function () {
-
-                                this.removeEventListener('click', arguments.callee);
-                                document.getElementById('delete').setAttribute("class", "modal")
-                                var conn = mysql.createConnection({
-                                    host: 'localhost',
-                                    user: 'CECYTEM',
-                                    password: '100%CECYTEM',
-                                    database: 'CECYTEM',
-                                    port: 3306
-                                })
-                                var q = conn.query('DELETE FROM TBL_MATERIA WHERE INT_MATERIA= ?', [element['INT_MATERIA']], function (error, result) {
-                                    if (error) {
-                                        throw error;
-                                    } else {
-                                        console.log("borrado")
-                                    }
+                            alertify.confirm("Eliminar Materia", "¿Desea eliminar la materia " + element['VCH_NOMBRE'] + "?",
+                                function () {
+                                    var con = mysql.createConnection({
+                                        host: 'localhost',
+                                        user: 'CECYTEM',
+                                        password: '100%CECYTEM',
+                                        database: 'CECYTEM',
+                                        port: 3306
+                                    })
+                                    var q = con.query('DELETE FROM TBL_MATERIA WHERE INT_MATERIA= ?', [element['INT_MATERIA']], function (error, result) {
+                                        if (error) {
+                                            throw error;
+                                        } else {
+                                            alertify.success("¡La materia se ha eliminado con éxito!");
+                                            setTimeout(function () {
+                                                location.href = "subject.html";
+                                            }, 3000);
+                                        }
+                                    });
+                                    con.end();
+                                },
+                                function () {
+                                    alertify.error('Se canceló la petición.');
                                 });
-                                conn.end();
+                            /*                            p = document.createElement('p')
+                                                        p.appendChild(document.createTextNode(element['VCH_NOMBRE']))
+                                                        document.getElementById('delete').setAttribute("class", "is-active modal")
+                                                        document.getElementById('text').appendChild(p)
+                                                        document.getElementById('delete-btn').addEventListener("click", function () {
 
-                            })
+                                                            this.removeEventListener('click', arguments.callee);
+                                                            document.getElementById('delete').setAttribute("class", "modal")
+                                                            var conn = mysql.createConnection({
+                                                                host: 'localhost',
+                                                                user: 'CECYTEM',
+                                                                password: '100%CECYTEM',
+                                                                database: 'CECYTEM',
+                                                                port: 3306
+                                                            })
+                                                            var q = conn.query('DELETE FROM TBL_MATERIA WHERE INT_MATERIA= ?', [element['INT_MATERIA']], function (error, result) {
+                                                                if (error) {
+                                                                    throw error;
+                                                                } else {
+                                                                    console.log("borrado")
+                                                                }
+                                                            });
+                                                            conn.end();
 
+                                                        })
+                            */
                         })
                         td.appendChild(span2)
                         tr.appendChild(td)
@@ -138,12 +164,57 @@ var app = new Vue({
                                     database: 'CECYTEM',
                                     port: 3306
                                 })
-                                var q = conn.query('UPDATE TBL_MATERIA SET VCH_NOMBRE= ?,INT_SEMESTRE=?,INT_HORAS_P=?,INT_HORAS_T=? WHERE INT_MATERIA= ?', [document.getElementById('materia').value, document.getElementById('semestre').value, document.getElementById('horap').value, document.getElementById('horat').value, element['INT_MATERIA']], function (error, result) {
+
+                                var materia = document.getElementById("materia").value,
+                                    semestre = document.getElementById("semestre").value,
+                                    horap = document.getElementById("horap").value,
+                                    horat = document.getElementById("horat").value;
+
+                                if (materia === "") {
+                                    alertify.error("Falta llenar el campo de MATERIA");
+                                    document.getElementById("materia").setAttribute("class", "input is-danger");
+                                    document.getElementById("errorMateria").setAttribute("style", "display: block");
+                                    return 0;
+                                } else {
+                                    document.getElementById("materia").setAttribute("class", "input is-success");
+                                    document.getElementById("errorMateria").setAttribute("style", "display: none");
+                                }
+                                if (semestre === "") {
+                                    alertify.error("Falta llenar el campo de SEMESTRE");
+                                    document.getElementById("semestre").setAttribute("class", "input is-danger");
+                                    document.getElementById("errorSemestre").setAttribute("style", "display: block");
+                                    return 0;
+                                } else {
+                                    document.getElementById("semestre").setAttribute("class", "input is-success");
+                                    document.getElementById("errorSemestre").setAttribute("style", "display: none");
+                                }
+                                if (horap === "") {
+                                    alertify.error("Falta llenar el campo de HORAS PRACTICAS");
+                                    document.getElementById("horap").setAttribute("class", "input is-danger");
+                                    document.getElementById("errorPra").setAttribute("style", "display: block");
+                                    return 0;
+                                } else {
+                                    document.getElementById("horap").setAttribute("class", "input is-success");
+                                    document.getElementById("errorPra").setAttribute("style", "display: none");
+                                }
+                                if (horat === "") {
+                                    alertify.error("Falta llenar el campo de MATERIA");
+                                    document.getElementById("horat").setAttribute("class", "input is-danger");
+                                    document.getElementById("errorTeo").setAttribute("style", "display: block");
+                                    return 0;
+                                } else {
+                                    document.getElementById("horat").setAttribute("class", "input is-success");
+                                    document.getElementById("errorTeo").setAttribute("style", "display: none");
+                                }
+
+                                var q = conn.query('UPDATE TBL_MATERIA SET VCH_NOMBRE= ?,INT_SEMESTRE=?,INT_HORAS_P=?,INT_HORAS_T=? WHERE INT_MATERIA= ?', [materia, semestre, horap, horat, element['INT_MATERIA']], function (error, result) {
                                     if (error) {
                                         throw error;
                                     } else {
-                                        console.log("actualizado")
-                                        location.reload();
+                                        alertify.success("¡Se actualizó la información de la materia!");
+                                        setTimeout(function () {
+                                            location.href = "subject.html";
+                                        }, 3000);
                                     }
                                 });
                                 conn.end();
@@ -169,7 +240,7 @@ var app = new Vue({
                 }
             });
 
-            connection.end();
+            con.end();
 
 
         }
